refactor(view-definition): ViewDefinitionEditor.tsx 分割 (#1145 Phase-4)
ViewDefinitionEditor.tsx 81KB / 1766 行 → 30KB / 744 行 (-62%、-51KB) に縮減。

責務分離:
- internal/Level1QueryEditor.tsx — Level 1 (Simple, sourceTableId)
- internal/Level2QueryEditor.tsx — Level 2 (Structured: from + joins + where/groupBy/having/orderBy)
- internal/Level3QueryEditor.tsx — Level 3 (Raw SQL + parameterRefs)
- internal/BasicInfoSection.tsx — Section 1 (id/name/kind/Level radio/maturity)
- internal/ColumnsSection.tsx — Section 3 (columns 編集テーブル + 列ごとの issue 表示)
- internal/SortDefaultsSection.tsx — Section 4 (sortDefaults)
- internal/FilterDefaultsSection.tsx — Section 5 (filterDefaults)
- internal/MiscSection.tsx — Section 6 (pageSize / groupBy)
- internal/IssueHints.tsx — 共通 issue inline 表示
- internal/useViewDefinitionTables.ts — useTablesForValidator / useTableOptions
- internal/viewDefinitionConstants.ts — FIELD_TYPE_OPTIONS / FILTER_OPERATORS / KIND_LABELS / JOIN_KIND_OPTIONS / LEVEL_LABELS / isBuiltinKind

親コンポーネントは 編集セッション / draft 管理 / state orchestration + 6 section 配置に専念。
列操作 (add/remove/move/setColRefTable/setColRefColumn) は colRefTableIds state を保持する
親側に残し、ColumnsSection に props で渡す (Phase-3 InspectorPanel と同パターン)。

test 追加:
- internal/view-definition-editors.test.tsx — 全 sub-editor / section に 21 件の
  rendering / 主要 callback test (FIELD_TYPE_OPTIONS / isBuiltinKind 含む)。
- 全 2243 → 2264 (+21) pass、tsc / lint 既存と同等。

Refs #1145
